# faster running sums