- again more unsafe :cry: 
- don't set buffer values twice

```
before
test filter::test::bench_box_filter                                     ... bench:   5,367,572 ns/iter (+/- 302,698)
test integralimage::test::bench_column_running_sum                      ... bench:       4,249 ns/iter (+/- 69)
test integralimage::test::bench_row_running_sum                         ... bench:       3,898 ns/iter (+/- 41)

after
test filter::test::bench_box_filter                                     ... bench:   3,812,161 ns/iter (+/- 30,126)
test integralimage::test::bench_column_running_sum                      ... bench:       1,126 ns/iter (+/- 87)
test integralimage::test::bench_row_running_sum                         ... bench:         838 ns/iter (+/- 14)
```